real toolbars blocking
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/trunk@4064 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ kadu/toolbutton.cpp
@@ -116,17 +116,21 @@ void ToolButton::mouseMoveEvent(QMouseEvent* e)
 void ToolButton::contextMenuEvent(QContextMenuEvent* e)
 {
 	kdebugf();
-	QPopupMenu* p = new QPopupMenu(this);
-	int label_menu_id =
-		p->insertItem(tr("Show text label"), this, SLOT(showTextLabelClicked()));
-	p->setItemChecked(label_menu_id, usesTextLabel());
-	p->insertItem(tr("Delete button"), this, SLOT(deleteButtonClicked()));
-	p->insertSeparator();
-	QPopupMenu* toolbar_menu = toolbar()->createContextMenu(p);
-	p->insertItem(tr("Toolbar menu"), toolbar_menu);
-	p->exec(QCursor::pos());
-	delete p;
-	e->accept();
+	if (DockArea::blocked())
+		e->ignore();
+	else
+	{
+		QPopupMenu* p = new QPopupMenu(this);
+		int label_menu_id = p->insertItem(tr("Show text label"), this, SLOT(showTextLabelClicked()));
+		p->setItemChecked(label_menu_id, usesTextLabel());
+		p->insertItem(tr("Delete button"), this, SLOT(deleteButtonClicked()));
+		p->insertSeparator();
+		QPopupMenu* toolbar_menu = toolbar()->createContextMenu(p);
+		p->insertItem(tr("Toolbar menu"), toolbar_menu);
+		p->exec(QCursor::pos() + QPoint(10, 10));
+		delete p;
+		e->accept();
+	}
 	kdebugf2();
 }
